# HW5/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Movie APP
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /movie/{title}:
    # our controller name
    x-swagger-router-controller: movie
    get:
      operationId : getOne
      description: get a movie
      # define the type of response for Success "200" and Error
      parameters:
        - name: title
          type: string
          in: path
          required: true
        - name: review
          type: boolean
          description: Check to see review
          in: query
          # schema:
          #     $ref: "#/definitions/CheckMovieReview"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetMovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId : update
      description: update a movie
      # define the parameters
      parameters:
        - name: title
          description: Movie title
          type: string
          in: path
          required: true
        - name: year
          description: Movie properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/AddNewMovieResponse"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AddNewMovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId : delete_movie
      description: delete a movie
      # define the parameters
      parameters:
        - name: title
          description: Movie title
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AddNewMovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  AddNewMovieResponse:
    required:
      - year
      - actor
    properties:
      year:
        type: number
      actor:
        type: array
        maximum: 3
        minimum: 3
        items:
          type: string
          maximum: 3
          minimum: 3
  # CheckMovieReview:
  #   properties:
  #     review:
  #       type: boolean
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GetMovieResponse:
    required:
      - year
      - actor
      - review
    properties:
      review:
        type: boolean
      year:
        type: number
      actor:
        type: array
        items:
          type: string
          maximum: 3
          minimum: 3
...